Fix non-linux builds
Adding build tags for linux and added stubs for other operating systems. This should ensure that code using it will always compile, but only work on linux.

I am unsure if FreeBSD and friends are expected to work, but for now it is "linux or nothing".
