+++ examples/smart_linux_test.go
